# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Collection": {
        "example": {
          "cover": "",
          "id": "col_example",
          "owner": "usr_example",
          "songs": [
            {
              "id": "song_example",
              "key": null,
              "nr": null
            }
          ],
          "title": "Sunday worship"
        },
        "properties": {
          "cover": {
            "description": "Cover art reference (client-resolved blob id or URL).",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "songs": {
            "items": {
              "$ref": "#/components/schemas/SongLink"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner",
          "title",
          "cover",
          "songs"
        ],
        "type": "object"
      },
...
      "CreateCollection": {
        "additionalProperties": false,
        "example": {
          "cover": "",
          "owner": "team_example_id",
          "songs": [
            {
              "id": "song_example",
              "key": null,
              "nr": null
            }
          ],
          "title": "Sunday worship"
        },
        "properties": {
          "cover": {
            "type": "string"
          },
          "owner": {
            "description": "Owning team id (`team` record id, same format as `Collection.owner` in responses). Omit to create under the caller's personal team.",
            "nullable": true,
            "type": "string"
          },
          "songs": {
            "items": {
              "$ref": "#/components/schemas/SongLink"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "cover",
          "songs"
        ],
        "type": "object"
      },
...
      "MoveOwner": {
        "additionalProperties": false,
        "description": "Target owning team for a library resource move (`owner` string matches GET responses).",
        "example": {
          "owner": "team_target_id"
        },
        "properties": {
          "owner": {
            "type": "string"
          }
        },
        "required": [
          "owner"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "[RFC 7807](https://www.rfc-editor.org/rfc/rfc7807) problem document (`application/problem+json`).\n\nCanonical error body for HTTP 4xx/5xx responses. Extension members include `code`.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code (extension member).",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable explanation.",
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "description": "Optional URI reference that identifies the specific occurrence.",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem class.",
            "type": "string"
          },
          "type": {
            "description": "URI reference identifying the problem type.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "SongLink": {
        "properties": {
          "id": {
            "description": "Song record id.",
            "type": "string"
          },
          "key": {
            "description": "Musical key for this slot (e.g. `G`, `Am`, `F#m`).",
            "nullable": true,
            "type": "string"
          },
          "nr": {
            "description": "Optional display position in the parent list (e.g. `1`, `2a`).",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Session cookie returned after a successful authentication flow",
        "in": "cookie",
        "name": "sso_session",
        "type": "apiKey"
      },
      "SessionToken": {
        "description": "Session override using `Authorization: Bearer <session>` header",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Worship Viewer API",
    "version": "2.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "operationId": "get_collections",
        "parameters": [
          {
            "description": "Zero-based page (default 0). `X-Total-Count` = filtered total before pagination; last page when `items.len() < page_size` or empty (`list-pagination.md`).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Items per page. Must be 1–500. Defaults to 50.",
            "example": 50,
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 500,
              "minimum": 1,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Full-text search query (title); uses text_search analyzer (stemming)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Collection"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Return all collections. `X-Total-Count` header contains the total number of matching collections."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid pagination parameters"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to fetch collections"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Collections"
        ]
      },
      "post": {
        "operationId": "create_collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            },
            "description": "Create a new collection. Optional request field `owner` is a team id (same format as `Collection.owner`); omit to create under the caller's personal team. Library edit access is required on the target team."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid collection payload"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Target team not found or caller cannot edit that team's library"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to create collection"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Collections"
        ]
      }
    },
...
    "/api/v1/collections/{id}/move": {
      "post": {
        "operationId": "move_collection",
        "parameters": [
          {
            "description": "Collection identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveOwner"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            },
            "description": "Collection moved to the target team, or unchanged when already owned by that team (idempotent)."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid `owner` team id"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Collection not found, target team not found, or caller lacks library write access on the current or destination team"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to move collection"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Collections"
        ]
      }
    },
    "/api/v1/collections/{id}/player": {
      "get": {
        "operationId": "get_collection_player",
        "parameters": [
          {
            "description": "Collection identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            },
            "description": "Return player metadata for a collection"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid collection identifier"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Collection not found"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No supported representation in Accept header"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to fetch collection player data"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Collections"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same origin as the web app (override per deployment).",
      "url": "/"
    },
    {
      "description": "Production deployment (public web app).",
      "url": "https://app.worshipviewer.com"
    }
  ],
...
}